InmateDetails Cache Refresh Improvements
Analyse of cache misses in production suggests that the refresh worker is stopping prematurely due to redis lock timeouts. This results in an indeterminate number of entries in the cache never being updated by the worker, resulting in user requests being slow as these trigger cache updates instead.

This commit helps address these issues as follows:

1. Adds logging and a sentry alert when the refresh worker ends premature. This includes information to help determine what an optimal lock timeout would be
2. Sorts the noms numbers to refresh randomly to ensure that even if the worker ends prematurely, we’re not just refreshing the same batch of noms numbers (this assumes the SQL query returns them in a consistnet order, that is indeterminate)
3. Adds a log when the inmate cache refresh worker completes to help future investigations