Add domain_events.schema_version column
This is used to track the schema version used for the ‘data’ property. The schema version
 will be specific to the corresponding ‘type’

This will be null for any domain events recorded before this concept was introduced and is optional. The domain event schema version used for a given domain event record is only required when the initial domain event schema changes in a way that we need to start tracking which schema a domain event uses.

For example, this change was added because we need to add some new mandatory fields to the cas1 booking cancelled domain event schema. We want to be able to easily determine if a domain event ‘data’ json has these fields when attempting to replay domain events so we know to ‘fix up’ the json to match the new schema.