Only retry specific status codes
ba71b6e990f51a60434c193fee1fb34ed2ff912f introduces retrying calls to API calls.

The intention was to only retry on specific status codes, but due to a combination of bad retry logic and a bad test, the code was retrying for all status codes. This commit fixes that.