Move toDomainEvent fun from entity to service
This toDomainEvent function is used to deserialize domain event json data into a Kotlin Instance. This commit moves the function from the DomainEventEntity to the CAS1 specific DomainEventService.

This has been done becase we need to add ‘schema version specific’ behaviour into this function. Adding this inside the DomainEventEntity was going to increase the size of an already large file. Also, because the function/entity was not managed by spring, it wouldn’t be possible to inject dependencies to use in this function, making it clunky to delegate the work to another class.

There some additional justifications for moving this function:

1. The function is only applicable to CAS1 domain events, where as the DomainEventEntity is used by CAS1, CAS2 & CAS3
2. DomainEventService is already responsible for serializing Kotlin Instances to JSON. This commit also makes it responsible for deserializing the JSON it serialized in the first place

In making this change it is no longer reified (and subsequently inline’d) as we need to call the function directly from DailyMetricsReportGenerator. This barely affects the size of the code and results in a more flexible functiond definition (reified/inline constraints functions)