Map CAS1 Booking Cancelled Event json v1 to v2
TODO: add a unit test for Cas1DomainEventMigrationService

This commit will map the JSON for a APPROVED_PREMISES_BOOKING_CANCELLED domain event persisted using the V1 JSON schema to the V2 JSON schema on-the-fly, whenever the domain event JSON is required. This mapping will add the two attributes ‘cancellationRecordedAt’ and ‘’cancelledAtDate.

There are two use cases for which this will occur:

1. When loading domain event details to render on the timeline (in the DomainEventDescriber)
2. When probation-integration retrieves the domain event json via the API

This change allows us to revert the short term fix introduces by ed66ab0, which amde the cancellationRecordedAt and cancelledAtDate attributes in the domain event schema optional to allow us to deserialize JSON into Kotlin Objects to render them on the timeline.

DomainEventServiceTest has been updated to leverage the new ‘schemaVersions’ attribute on the DomainEventType enumeration, testing the retrieval of all DomainEventType/SchemaVersion combinations.