Map CAS1 Booking Cancelled Event json v1 to v2
This commit will adapt the V1 JSON persisted for a APPROVED_PREMISES_BOOKING_CANCELLED domain event to be compatible V2 JSON schema on-the-fly, whenever the domain event JSON is required. This adaption will add the two attributes ‘cancellationRecordedAt’ and ‘cancelledAtDate’

This is required for the following scenarios:

1. When loading domain event details to render on the timeline (in the DomainEventDescriber)
2. When probation-integration retrieves the domain event json via the API

This change allows us to revert the short term fix introduced by ed66ab0, which made the cancellationRecordedAt and cancelledAtDate attributes in the domain event schema optional to allow us to deserialize JSON into Kotlin Objects to render them on the timeline.

DomainEventServiceTest has been updated to leverage the new ‘schemaVersions’ attribute on the DomainEventType enumeration, testing the retrieval of all DomainEventType/SchemaVersion combinations.

This commit also adds a Mark Down file providing guidance on how to do this in the future.