Refactor intellij optimisations in gradle
This commit changes how we disable open api generation when running tests in intellij.

Previously the task determined whehter it should run at the point of execution. This is now determined when the gradle file is parsed. This helps us move towards the possibility of enabling the configuration cache, which doesn’t allow the project instance to be accessed when a task is executed.