The MDC Filter was unnecessarily loading the user
The MDCHandlerInterceptor only required access to the delius username, but the function it used to access this on UserService was loading the whole user from the database, triggering 3 SQL queries.

This was also operating outside of our exception handler, meaning we weren’t returning a 503 if we had database connection issues.