OOSB update shouldn’t check for booking conflicts
Going forward we will not maintain the concept of a booking being assigned to a bed. Therefore, we should not check for such conflicts when updating an out of service bed (OOSB). Note that this is already the case when creating an OOSB.

This change is required to avoid unexpected errors in production related to bookings that _have_ been allocated to a bed in the legacy version of match and manage.