Convert eager load to lazy for performance
This commit updates some @ManyToOne and @OneToOne relationships to use eager loading to reduce the number of unneccessary database calls in production.

Placement Date Entity - the relationships now marked as eager were only used in the hashcode/equals generation, which was unneccessary given that the ID uniquely identifies an instance

Placement Request Entity - the assessment property is not typically required but eagerly loaded. This is now lazy.

The change to placement request entity highlighted that findByIdOrNull on spring data repositories can return a hibernate proxy if the entity isn’t already loaded into the session. Change to the eager loading behaviour meant an assessment was no longer loading into a session by default during an integration test, leading to a failure when doing an ‘is’ comparison on the assessment. I’ve reviewed all usages of findByIdOrNull and none other would be affected by this unexpected behaviour other than the one i’ve fixed.