Clean up domain events with incorrect date format.
Recently a new domain event type 'CAS3_ASSESSMENT_UPDATED' was added. Initially the domain event json was using the incorrect format dates. This has since been corrected in the code.
This commit adds a migration job to remove the small handful of domain events that have been created with the incorrect formatting