Link PlacementRequest to CAS1SpaceBooking
This commit introduces a OneToMany relationship between placement requests and space bookings.

It’s worth noting that the relationship between placement requests and bookings is ManyToOne. This is technically incorrect and should be OneToOne. If a booking is cancelled and then subsequently replaced by another booking, the relationship to the cancelled booking to its placement request is lost. Because this information may be of value, we do not mimick that behaviour for Space Bookings and instead maintain a complete list of space bookings, cancelled or otherwise