Consider Space Bookings as Withdrawable
This commit adds logic to consider Space Bookings when building the WithdrawableTree and consequently returns Withdrawables of type ‘SpaceBooking’ where relevant.

This uses the same look as Bookings to determine if a Space Booking is withdrawable, with the following exceptions: We do not check the arrival state in delius because Arrival information will be managed in CAS1 for Space Bookings.

Note that this commit does not add functionality to request withdrawal of a SpaceBooking. That will follow in a subsequent change.