Create Cas1ApplicationStatusService
Initially we used Entity Listeners to update the Application’s status when other entities change. This approach wasn’t ideal because it created fragmented logic that was hard to reason upon and understand the order in which it will be applied.

As part of the spring boot 3 upgrade we could no longer use the Entity Listeners (see 0b61286). To minimise change we converted these into simple spring components that the code was explictly call before persisting changes to Entities.

This commit moves us towards the target state of having a single service reponsible for managing application status changes, giving us a more cohesive component for all application status changes.

This commit also removes the BookingListener and replaces it wtih a call to Cas1ApplicationStatusService