removed use of community api in UserService
This PR removes the feature flag for using the ap-and-delius integration - it was set to true for all environments. It also removes the use of the community-api to get staff user details in the UserService, instead getting them from the AP and Delius integration. It also renamed a bunch of things to remove the 'communityApi' part, to a more generic 'byApi'.