Make Ap Area Mandatory on submit CAS1 Application
On submission we must have the AP Area set on the application.

The code made it possible for this to be null if the UI provided an invalid UUID.

This code will now throw an error if this happens.

I’ve double chceked we have no submitted applications in production where ap area isn’t set using

```
select count(*) from applications a
inner join approved_premises_applications apa on a.id = apa.id
where a.submitted_at is not null and apa.ap_area_id is null;
```