Add CRU Management Area to user
THis commit adds two new fields to the users table:

* cas1_cru_management_area_id
* cas1_cru_management_area_override_id

The former will always be updated when ap area is updated, using the ap areas default management area value.

This commit also backfills the cas1_cru_management_area_id value for existing users.

The later will be configurable via an API call to be added in a subsquent commit.