Remove legacy FemaleAssessmentRule
This rule was used to automatically assign assessments for ‘womens applications’ to an assessor with the ‘WOMENS’ qualification. This is going to be replaced by the new women’s estate logic. Furthermore, there is no user in any environment with the ‘WOMENS’ qualification, and there is no option to add this qualification on the UI. Therefore, this is safe to remove.