Add auto allocation config to CRU Management Area
This commit adds an ‘assessmentAutoAllocationUsername’ property the Cas1CruManagementAreaEntity, to be used for auto allocation. This field isn’t currently in use, this will follow in a subsequent commit.

This commit also adds a seed job that can be used to update a CRU Managment Areas email and auto allocation configuration.

The default values for assessmentAutoAllocationUsername in local/dev/test are set to match their helm equivalents