Add simple timeline description for space bookings
Whilst space bookings cannot currently be directly withdrawn on the UI, they can be withdrawn indirectly by withdrawing the associated application.

This leads to a domain event entry for space booking withdrawal but the timeline describer trys to render this using a regular booking entity, which fails because a regular booking with that ID doesn’t exist.

This PR adds a simple timeline description for space booking withdrawal to resolve the error. This will be replaced with a complete message in a subsequent commit.