Remove StaffUserDetails from integration tests.
This PR removes StaffUserDetails from the integration tests, as well as removing the factory. The StaffUserDetails class will need to be removed in a later PR, as the mocked calls to get staff by code still use it. It also updates the delius name to use all three names, that was inadvertantly removed in a previous commit.