Add bootRunDebug task, removing hardcoded profile
We currently use bootRunLocal to start the API from ap-tools. This task hardcodes the spring profile being used to ‘local’. To support running tests against upstream services, we want to be able to run against the ‘dev’ profile. Therefore, this commit introduces a new ‘bootRunDebug’ task that is identical to ‘bootRunLocal’ but removes the hardcoded profile.

The calling ap-tools code already sets the active profile in the env var, so this will not breaking existing behaviour