Add Space Booking Day Planner Logic
This commits a SpaceBookingDayPlanner service. Given a list of rooms and space bookings, this will attempt to optimally assign a a space booking to each room, where possible.

Subsequent commits will make use of this planner to plan occupancy over a period of time for a given premise, which will then be used for occupancy calculation.