Space Day Planner Rendering Improvements
Before this commit when rendering the day plan results we would indicate if the room characteristic matched a required booking characteristic using a + or - symbol.

This commit changes this output to list out all characteristics of the room or required of the booking, adding a ‘r’ to indicate the room has the characteristic and ‘b’ to indicate the booking requires the characteristic. This then allows us to show single room characteristics in the output, a ‘special’ characteristic that only exists on bookings