Add multi-day Space Planner
This commit adds a `SpacePlanner` class that can be used to plan  `SpaceBooking` to `Bed` assignments for a configurable period of days.

This is not currently in use by any APIs. In a subsequent commit an API will be added to invoke the planner a given premise and period of time to verify the planning output against space booking data in test environments.