Offline app support in CAS1 Space Bookings
We have several bookings linked to `offline_applications` with space bookings arriving and/or departing in the future. Therefore, these need supporting.

This commit updates the cas1_space_bookings datamodel to support space bookings being linked to an `offline_application` instead of an `approved_premises_application`

This requires update to email and domain event logic, which mimics the behaviour currently used for legacy `bookings`. Note that withdrawal emails are not sent for bookings (or space bookings) linked to offline bookings. A ticket APS-1493 has been created to address that.

Note that this commit is not sufficient to fully support the migration of bookings linked to offline applications to space bookings because placement requests will also not be defined for such bookings. That will be addressed in subsequent commits, along with space booking migration tests for this scenario.