Limit space planning to specific characteristics
Whilst rooms have many characteristics, only a handful can be considered mandatory on a `placement_request`. Given this, this commit limits the characteristics considered for planning to those that will have an impact on planning. This considerably simplifies the planning output as irrelevant characteristics on the room are no longer shown.