Add method to quickly recycle CasResult error
Sometimes in a service we want to return a `CasResult.Error` ‘as is’ back to the caller. Before this commit we would have to ‘rebuild’ the `CasResult` specific type to make the generic arguments happy, despite the SuccessType generic not being used anyway in the `CasResult.Error` hierarchy. This was verbose and forced us to reassert information that was already in the original error, typically loosing context.

This commit adds a `reviseType` function to `CasResult.Error` to allow us to directly use the `CasResult.Error` instance by forcefully changing its generic.

It also makes use of the function in `PlacementApplicationService` to provide a working example.