Use booking ID for migrated space bookings
When converting a booking into a space booking we want to use the existing bookings ID. This is ‘safe’ as we use UUID for IDs.

This will ensure that any domain event triggered against the space booking will still work, as the booking id is stored against the corresponding referral in delius.