Space bookings departing < June 24 are historic
As part of the migration to space bookings we have several older bookings for which we cannot easily backfill arrival and departure data. To minimise the amount of manual effort to cleanse such data we have decided to assume that any booking with an expected departure before June ’24 as having been completed, and as such treated as a ‘historic’ booking when listing space bookings.