Split space booking arrival & departure date/time
This commit splits the arrival and departure datetime fields into individual date and time fields in the persisted data model

This change is required to handle delius imports which only provide the date of an arrival and/or departure. Furthermore this structure better reflects how the date and time are split on the UI.

Both date and time are captured as ‘wall clock’ values

The API is unaffected by this change. A subsequent commit will expose these individual values as appropriate on the API and deprecate the existing date/time API values.