Add seperate date/time fields to CAS1 arrival API
This commit adds separate date and time fields to the space bookings arrival and departure APIs. The existing datetime fields have been deprecated and will be removed once the UI has been updated to use the new fields. At this point the new fields will be marked as mandatory on the API.