Fix Space Booking Migration Reason Code Resolution
This commit fixes an issue caused by there being multiple entries in the departure reason code table with the same legacy delius code. One entry is a enabled, the other is not. This results in a `Query did not return a unique result: 2 results were returned` error.

This commit fixes this issue by getting a list of matching reason codes, taking the enabled one if it exists.