Booking to Space Booking - remove old DateChanges
When a Booking is removed as part of the Space Booking migration, any DateChange entries related to the old booking should also be removed. Otherwise a FK reference error will occur. Note that date changes are captured as domain events, so this data will still be available for space bookings.