Improve Risk Datasource Error Message
Before this commit the `OffenderRisksDataSource` recorded a RuntimeException in sentry if risks couldn’t be loaded. The error message implied that processing of the thread had failed/errored because of this error, but that’s typically not the case as the calling code can continue if risks arne’t known (and will error itself if it can’t).