Remove StaffDetail.staffIdentifier
This change is being made in preparation for StaffDetail.staffIdentifier to be removed from ap-and-delius responses.

We only use this field to populate `StaffMember.staffIdentifier` on domain events, but this value is not used by the consumer. Therefore, it’s temporarily being set to -1, until the consumer is ready for us to remove `StaffMember.staffIdentifier` completely from the domain events.