Add seed job to update space booking fields
This commit adds a seed job that can update a space bookings event number, and associated characteristics.

An event number can only be updated if the space booking is linked to an offline application for which no event number is assigned. In all other scenarios the `Cas1UpdateEventNumberSeedJob` should be use to update the application’s event number which will cascade to the space booking (change to be made in a subsequent commit).