Emit CAS1 arrival and departure events
Previously we disabled emitting CAS1 arrival and departure domain events because there was functionality available on the user interface to record arrivals and departures, but we did no want this being sent to delius as the functonality was incomplete.

There is no longer ‘record arrival/departure’ functionality available to users in prod so it’s safe to re-enable these domain events to support testing of space booking management.