Replace redundant get CAS1 app test
This commit removes the the test `Get single application returns 403 when caller not in a managing team and user is not one of roles WORKFLOW_MANAGER, ASSESSOR, MATCHER, MANAGER`

The description did not match the implemented logic. Instead, it was testing the case where the offender was LAO and the caller didn’t have permission to access the application. This is already covered by `Get single LAO application for non creator, no LAO Access, no LAO Qualification returns 403`

The test has been replaced with a test to ensure that other users can access applications if the offender isn’t LAO