Fix Integration Tests wrt. LAO users
The CAS1 ‘can view application’ logic was using the response from getOffender() to determine if the offender could be viewed by the user. This usage was too broad because a NotFound response was treated as an LAO restriction, which isn’t the case.

This commit fixes some Integration Tests that relied on this behaviour to instead generate an actual restricted scenario. Use of the getOffender() function will be removed in a subsequent commit.