Get placement request 403 if cant access LAO
Before this commit, it was possible to retrieve placement request details even if you did not have access to the Offender details due to the offender being an LAO. Whilst the Person info was redacted in this case, application detail was not. Note that the UI wasn’t showing any of the sensitive application information, but regardless we should not be returning it from the API.

Given that the UI doesn’t link to placemnt requests if the user doesn’t have access to an LAO offender, this isn’t a breaking change.