Capture recordedBy on CAS1 arrival domain events
Domain event schema version control is used to identify older domain events that don’t have a ‘recordedBy’ element in the JSON and backfill the value with the domain event’s triggered by user, if available.

The primary purpose of this migration is to make v1 domain events schema valid. `recordedBy` is not used to render the timeline, and these old domain events will not be consumed externally, so the imperfect nature of the back fill is acceptable.