Remove unused availability info from space search
When the Space Search Service was originally written, it was expected that the results would include availability for each returned premise. This is no longer the case and is managed by a separate API.

This commit removes the ‘placeholder’ code added to provide capacity, simplifying the response from the space search service.

It doesn’t make any functional changes to the code