Add ‘apType’ filter option to space search
This commit deprecates the existing `apTypes` filter option for space search, replacing it instead with a single `apType` filter option. This helps simplify the code and reflects how the search is being used by the UI.

This commit also makes the query more efficient, removing subsqueries and instead using group aggregations and arrays to assert on characteristics.