Add approved_premises.full_address
The AP site surveys provide the address on a single line (excluding postcode)

This commit introduces a new persisted field `approved_premises.full_address` that should be used going forward instead of the individual address line or town fields. It updates the space search API to backfill the full address from the indiviual fields ‘on the fly’ to account for premises that have not yet been updated from the site surveys.